# openapi.yaml
openapi: 3.0.0
paths:
...
  /residents/{residentUid}/bath-records:
    get:
      operationId: getBathRecords
      parameters:
        - name: residentUid
          required: true
          in: path
          description: 利用者UID
          schema:
            type: string
      responses:
        "200":
          description: 入浴記録一覧の取得に成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BathRecordListResponseDto"
      security:
        - JWT-auth: []
      summary: 利用者の入浴記録一覧を取得
      tags: &a6
        - bath-records
    post:
      operationId: createBathRecord
      parameters:
        - name: residentUid
          required: true
          in: path
          description: 利用者UID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BathRecordCreateInputDto"
      responses:
        "201":
          description: 入浴記録の作成に成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BathRecordDto"
      security:
        - JWT-auth: []
      summary: 入浴記録を作成
      tags: *a6
  /residents/{residentUid}/bath-records/{uid}:
    patch:
      operationId: updateBathRecord
      parameters:
        - name: uid
          required: true
          in: path
          description: 入浴記録UID
          schema:
            type: string
        - name: residentUid
          required: true
          in: path
          description: 利用者UID
          schema: {}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BathRecordUpdateInputDto"
      responses:
        "200":
          description: 入浴記録の更新に成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BathRecordDto"
      security:
        - JWT-auth: []
      summary: 入浴記録を更新
      tags: *a6
    delete:
      operationId: deleteBathRecord
      parameters:
        - name: uid
          required: true
          in: path
          description: 入浴記録UID
          schema:
            type: string
        - name: residentUid
          required: true
          in: path
          description: 利用者UID
          schema: {}
      responses:
        "200":
          description: 入浴記録の削除に成功
      security:
        - JWT-auth: []
      summary: 入浴記録を削除
      tags: *a6
  /residents/{residentUid}/bath-records/{uid}/transcription:
    get:
      operationId: getBathRecordTranscription
      parameters:
        - name: uid
          required: true
          in: path
          description: 入浴記録UID
          schema:
            type: string
        - name: residentUid
          required: true
          in: path
          description: 利用者UID
          schema: {}
      responses:
        "200":
          description: 入浴記録の文字起こし取得に成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TranscriptionDto"
      security:
        - JWT-auth: []
      summary: 入浴記録の文字起こしを取得
      tags: *a6
    patch:
      operationId: appendBathRecordTranscription
      parameters:
        - name: uid
          required: true
          in: path
          description: 入浴記録UID
          schema:
            type: string
        - name: residentUid
          required: true
          in: path
          description: 利用者UID
          schema: {}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TranscriptionInputDto"
      responses:
        "200":
          description: 入浴記録の文字起こし追記に成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TranscriptionDto"
      security:
        - JWT-auth: []
      summary: 入浴記録の文字起こしを追記
      tags: *a6
    put:
      operationId: updateBathRecordTranscription
      parameters:
        - name: uid
          required: true
          in: path
          description: 入浴記録UID
          schema:
            type: string
        - name: residentUid
          required: true
          in: path
          description: 利用者UID
          schema: {}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TranscriptionInputDto"
      responses:
        "200":
          description: 入浴記録の文字起こし置換に成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TranscriptionDto"
      security:
        - JWT-auth: []
      summary: 入浴記録の文字起こしを置換
      tags: *a6
    delete:
      operationId: deleteBathRecordTranscription
      parameters:
        - name: uid
          required: true
          in: path
          description: 入浴記録UID
          schema:
            type: string
        - name: residentUid
          required: true
          in: path
          description: 利用者UID
          schema: {}
      responses:
        "200":
          description: 入浴記録の文字起こし削除に成功
      security:
        - JWT-auth: []
      summary: 入浴記録の文字起こしを削除
      tags: *a6
  /residents/{residentUid}/bath-records/{uid}/extract:
    post:
      operationId: extractBathRecord
      parameters:
        - name: uid
          required: true
          in: path
          description: 入浴記録UID
          schema:
            type: string
        - name: residentUid
          required: true
          in: path
          description: 利用者UID
          schema: {}
      responses:
        "200":
          description: 入浴記録からの情報抽出に成功
          content:
            application/json:
              schema:
                type: string
      security:
        - JWT-auth: []
      summary: 入浴記録から情報を抽出
      tags: *a6
...
  /assessments/{uid}/transcription:
    get:
      operationId: getTranscription
      parameters:
        - name: uid
          required: true
          in: path
          description: アセスメントUID
          schema:
            type: string
      responses:
        "200":
          description: アセスメントの文字起こし取得に成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TranscriptionDto"
      security:
        - JWT-auth: []
      summary: アセスメントの文字起こしを取得
      tags: *a10
    patch:
      operationId: appendTranscription
      parameters:
        - name: uid
          required: true
          in: path
          description: アセスメントUID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TranscriptionInputDto"
      responses:
        "200":
          description: アセスメントの文字起こし追記に成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TranscriptionDto"
      security:
        - JWT-auth: []
      summary: アセスメントの文字起こしを追記
      tags: *a10
    put:
      operationId: updateTranscription
      parameters:
        - name: uid
          required: true
          in: path
          description: アセスメントUID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TranscriptionInputDto"
      responses:
        "200":
          description: アセスメントの文字起こし置換に成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TranscriptionDto"
      security:
        - JWT-auth: []
      summary: アセスメントの文字起こしを置換
      tags: *a10
    delete:
      operationId: deleteTranscription
      parameters:
        - name: uid
          required: true
          in: path
          description: アセスメントUID
          schema:
            type: string
      responses:
        "200":
          description: アセスメントの文字起こし削除に成功
      security:
        - JWT-auth: []
      summary: アセスメントの文字起こしを削除
      tags: *a10
...
info:
  title: Nagara Care API
  description: ""
  version: "1.0"
  contact: {}
...
servers: []
components:
  securitySchemes:
    JWT-auth:
      scheme: bearer
      bearerFormat: JWT
      type: http
      name: JWT
      description: Enter JWT token
      in: header
  schemas:
...
    TranscriptionDto:
      type: object
      properties:
        transcription:
          type: string
          description: 文字起こしテキスト
      required:
        - transcription
    TranscriptionInputDto:
      type: object
      properties:
        transcription:
          type: string
          description: 文字起こしテキスト
      required:
        - transcription
    BathRecordDto:
      type: object
      properties:
        uid:
          type: string
          description: UID
        tenantUid:
          type: string
          description: テナントUID
        caregiverUid:
          type: string
          description: 介護者UID
        residentUid:
          type: string
          description: 利用者UID
        recordedAt:
          format: date-time
          type: string
          description: 記録時刻
        notes:
          type: string
          description: メモ
        bathMethod:
          type: string
          description: 入浴方法
      required:
        - uid
        - tenantUid
        - caregiverUid
        - residentUid
        - recordedAt
        - notes
        - bathMethod
    BathRecordListResponseDto:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/BathRecordDto"
        total:
          type: number
          description: 総件数
      required:
        - items
        - total
    BathRecordCreateInputDto:
      type: object
      properties:
        recordedAt:
          format: date-time
          type: string
          description: 記録時刻
          example: 2024-03-20T12:00:00Z
        notes:
          type: string
          description: メモ
        bathMethod:
          type: string
          description: 入浴方法
          example: 一般浴
      required:
        - recordedAt
        - bathMethod
    BathRecordUpdateInputDto:
      type: object
      properties:
        recordedAt:
          format: date-time
          type: string
          description: 記録時刻
          example: 2024-03-20T12:00:00Z
        notes:
          type: string
          description: メモ
        bathMethod:
          type: string
          description: 入浴方法
          example: 一般浴
      required:
        - recordedAt
        - bathMethod
...